fix type of obj_affine_set param in new asm! syntax, don't use asm! in non-agb build in dma.rs

+++ src/bios.rs
@@ -511,7 +511,7 @@ pub fn obj_affine_set(src: *const ObjAffineSetParams, dest: usize, num_calc: u32
           in("r0") src,
           in("r1") dest,
           in("r2") num_calc,
-          in("r3") offset,
+          in("r3") offset as u32,
       );
     }
   }
